// FRONTEND/src/types/appData.ts
export interface TemporalCoverage {
  start_date: string | null;
  end_date: string | null;
}

interface LocationData {
  west_bound_longitude: `${number}` | null;
  east_bound_longitude: `${number}` | null;
  north_bound_longitude: `${number}` | null;
  south_bound_longitude: `${number}` | null;
  mean_latitude: `${number}` | null;
  mean_longitude: `${number}` | null;
}

export interface IDataset {
  id: IDatasetID;
  location_data: LocationData;
  doi: IDoi;
  dataset_publication_date?: IPublicationDate;
  temporal_coverage: TemporalCoverage | null;
  authors?: IAuthor[];
  providers?: IProvider[];
  dataset_title: ITitle;
  organization: "PANGAEA" | "Earth Observatory";
}

//Dataset attributes
export type IDatasetID =
  | `Dataset/${IDatasetKEY}`
  | `STACCollection/${IDatasetKEY}`;
export type IDatasetKEY = number | string;
export type IAbstract = string | null;
export type ITitle = string | null;
export type IAuthor = string | null;
export type IProvider = string | null;
export type IDoi = string | null;
export type IPublicationDate = string | null;